drawer refactor
- refactor the code and state for the drawer a bit to reduce ugliness
- move buttons outside of the settings panel
