+++ src/components/layout/Nav.jsx
@@ -1,6 +1,7 @@
 import React, { useState } from 'react'
 import { Dialog, Snackbar } from '@material-ui/core'
 import { Alert } from '@material-ui/lab'
+import shallow from 'zustand/shallow'
 
 import Utility from '@services/Utility'
 import useStyles from '@hooks/useStyles'
@@ -23,7 +24,6 @@ export default function Nav({
   map,
   setManualParams,
   Icons,
-  config,
   setWebhookMode,
   webhookMode,
   settings,
@@ -36,45 +36,43 @@ export default function Nav({
   isTablet,
 }) {
   const classes = useStyles()
-  const { perms } = useStatic((state) => state.auth)
-  const webhookAlert = useStatic((state) => state.webhookAlert)
-  const setWebhookAlert = useStatic((state) => state.setWebhookAlert)
   const {
-    map: { enableTutorial, messageOfTheDay, donationPage },
-  } = useStatic((state) => state.config)
-  const userProfile = useStatic((state) => state.userProfile)
-  const setUserProfile = useStatic((state) => state.setUserProfile)
-  const feedback = useStatic((state) => state.feedback)
-  const setFeedback = useStatic((state) => state.setFeedback)
-  const resetFilters = useStatic((state) => state.resetFilters)
-  const setResetFilters = useStatic((state) => state.setResetFilters)
+    auth: { perms },
+    setWebhookAlert,
+    config,
+    setUserProfile,
+    setFeedback,
+    setResetFilters,
+  } = useStatic.getState()
+  const { setFilters, setUserSettings, setTutorial, setMotdIndex } =
+    useStore.getState()
 
-  const filters = useStore((state) => state.filters)
-  const setFilters = useStore((state) => state.setFilters)
-  const userSettings = useStore((state) => state.userSettings)
-  const setUserSettings = useStore((state) => state.setUserSettings)
-  const tutorial = useStore((state) => state.tutorial)
-  const setTutorial = useStore((state) => state.setTutorial)
-  const motdIndex = useStore((state) => state.motdIndex)
-  const setMotdIndex = useStore((s) => s.setMotdIndex)
+  const { webhookAlert, userProfile, feedback, resetFilters } = useStatic(
+    (s) => s,
+    shallow,
+  )
+  const { filters, userSettings, tutorial, motdIndex } = useStore(
+    (s) => s,
+    shallow,
+  )
 
   const [drawer, setDrawer] = useState(false)
+  const [donorPage, setDonorPage] = useState(false)
   const [dialog, setDialog] = useState({
     open: false,
     category: '',
     type: '',
   })
   const [motd, setMotd] = useState(
-    messageOfTheDay.components?.length &&
-      (messageOfTheDay.index > motdIndex ||
-        messageOfTheDay.settings.permanent) &&
+    config.map.messageOfTheDay.components?.length &&
+      (config.map.messageOfTheDay.index > motdIndex ||
+        config.map.messageOfTheDay.settings.permanent) &&
       ((perms.donor
-        ? messageOfTheDay.settings.donorOnly
-        : messageOfTheDay.settings.freeloaderOnly) ||
-        (!messageOfTheDay.settings.donorOnly &&
-          !messageOfTheDay.settings.freeloaderOnly)),
+        ? config.map.messageOfTheDay.settings.donorOnly
+        : config.map.messageOfTheDay.settings.freeloaderOnly) ||
+        (!config.map.messageOfTheDay.settings.donorOnly &&
+          !config.map.messageOfTheDay.settings.freeloaderOnly)),
   )
-  const [donorPage, setDonorPage] = useState(false)
 
   const toggleDrawer = (open) => (event) => {
     if (
@@ -87,8 +85,8 @@ export default function Nav({
   }
 
   const handleMotdClose = () => {
-    if (!messageOfTheDay.settings.permanent) {
-      setMotdIndex(messageOfTheDay.index)
+    if (!config.map.messageOfTheDay.settings.permanent) {
+      setMotdIndex(config.map.messageOfTheDay.index)
     }
     setMotd(false)
   }
@@ -124,18 +122,14 @@ export default function Nav({
         <Sidebar
           drawer={drawer}
           toggleDrawer={toggleDrawer}
-          filters={filters}
-          setFilters={setFilters}
           toggleDialog={toggleDialog}
-          Icons={Icons}
         />
       ) : (
         <FloatingBtn
           toggleDrawer={toggleDrawer}
           toggleDialog={toggleDialog}
-          safeSearch={config.searchable}
+          safeSearch={config.map.searchable}
           isMobile={isMobile}
-          perms={perms}
           webhooks={webhooks}
           webhookMode={webhookMode}
           setWebhookMode={setWebhookMode}
@@ -144,7 +138,7 @@ export default function Nav({
           scanZoneMode={scanZoneMode}
           setScanZoneMode={setScanZoneMode}
           settings={settings}
-          donationPage={donationPage}
+          donationPage={config.map.donationPage}
           setDonorPage={setDonorPage}
           setUserProfile={setUserProfile}
         />
@@ -162,7 +156,7 @@ export default function Nav({
         />
       </Dialog>
       <Dialog
-        open={tutorial && enableTutorial}
+        open={tutorial && config.map.enableTutorial}
         fullScreen={isMobile}
         maxWidth="xs"
         onClose={() => setTutorial(false)}
@@ -221,14 +215,14 @@ export default function Nav({
         onClose={handleMotdClose}
       >
         <Motd
-          motd={messageOfTheDay}
+          motd={config.map.messageOfTheDay}
           perms={perms}
           handleMotdClose={handleMotdClose}
         />
       </Dialog>
       <Dialog open={donorPage} onClose={() => setDonorPage(false)}>
         <DonorPage
-          donorPage={donationPage}
+          donorPage={config.map.donationPage}
           handleDonorClose={() => setDonorPage(false)}
         />
       </Dialog>
